Langchain based Chatbot
Using The Langchain classes and RAG method with LangGraph to retrieve the results/answers from LLM.
The UI is built using Streamlit.
Note : The model used for embedding is adav3 and generation (answer) is gpt-3.5-turbo-instruct.  The model, vector db and prompt model can be dynamically passed :)
This works for any pdf.